refactor: rename Users to Accounts
